feat: 무한스크롤 구현 및 기타 설정 변경 (#71)
* fix: 프로필 사진이 없으면 기본 이미지 렌더링

* fix: Header 컴포넌트 고정 오류 해결

* fix: transition 설정 변경

* fix: 좋아요 페이지 정렬 api 요청 오류 해결

* fix: 기본 이미지 원 형태로 변경

* fix: transfrom 재설정

* fix: 좋아요 개수 0인 상태에서 좋아요 취소되는 오류 해결

* feat: 무한스크롤 훅 생성 및 적용

* fix: 무한스크롤 기능 중 observer disconnect 적용

* fix: 좋아요 개수 상태값 적용 오류 해결

* fix: 상태값 totalPages로 통일

+++ client/src/components/common/card/review_card/CardFront.jsx
@@ -9,7 +9,8 @@ import likesIcon from '../../../../assets/icons/heart_filled.png';
 import DropBoxButton from '../../button/DropBoxButton';
 import reviewApi from '../../../../api/reviewApi';
 import { useNavigate } from 'react-router-dom';
-import basic from '../.././../../assets/basic.jpg';
+import basic from '../../../../assets/basic.jpg';
+import basicProfileImage from '../../../../assets/user.png';
 
 export default function CardFront({
   review,
@@ -33,7 +34,7 @@ export default function CardFront({
   const userId = useSelector((state) => state.auth.memberId);
   const modalRef = useRef(null);
 
-  const navigate = useNavigate();
+  const navigate = useNavigate();  
 
   useEffect(() => {
     setIsLiked(like);
@@ -55,6 +56,8 @@ export default function CardFront({
         await reviewApi.postReviewLikes(id);
         setReviewLikeCount((prev) => prev + 1);
       } else {
+        if (reviewLikeCount === 0) return;
+
         await reviewApi.deleteReviewLikes(id);
         setReviewLikeCount((prev) => prev - 1);
       }
@@ -121,7 +124,7 @@ export default function CardFront({
       <div className={styles.bottomWrapper}>
         <div className={clsx(styles.flexBox, styles.justifyStart)} onClick={clickProfileHandler}>
           <div className={styles.profileImageWrapper}>
-            <img src={profileImageUrl} alt="profileImage" />
+            <img src={profileImageUrl || basicProfileImage} alt="profileImage" />
           </div>
           <div className={styles.nickname}>{nickname}</div>
         </div>
